Combining existing work
Created a new scene (CombinedScene) to combine existing work.
Zoomed out the camera to give a better view.
Reworked player character and player objects to be 2D and use 2D components.
Made my esteemed co-developer's flowers child objects of the boxes they are placed in, altered tags and created new colliders allowing them to be picked up by the player and moved into different flower beds.
Pests are now destroyed when the player interacts with them holding pesticide.
There is currently an issue with dropping a flower and picking it up again. When this is attempted, the player is hidden from view and then appears above a flower bed when they try to move. They then cannot pick up the plant. I am currently unsure of the cause of this.